Fix clippy stuff

+++ src/lib.rs
@@ -136,7 +136,7 @@ pub fn parse_and_resolve_names(flags: BTreeSet<Flag>, files: Vec<String>) {
                 panic!("ERR: {} {:?}", arg, e);
             }
             Ok(src) => {
-                let l = lexer::lex(&src, ast::Fi(i as usize));
+                let l = lexer::lex(&src, ast::Fi(i));
                 let mut p = parser::P::new(&l, &names);
                 if let Some(m) = parser::module(&mut p) {
                     let header = m.0.clone()?;
@@ -317,7 +317,7 @@ pub fn parse_modules(flags: BTreeSet<Flag>, files: Vec<String>) {
             Ok(src) => {
                 use std::io::BufWriter;
 
-                let l = lexer::lex(&src, ast::Fi(i as usize));
+                let l = lexer::lex(&src, ast::Fi(i));
                 let n = DashMap::new();
                 let mut p = parser::P::new(&l, &n);
